add Insert(pos, iter_first ,iter_last) and Concat for data modification.

+++ Examples/en/quickstart_stable.cpp
@@ -227,7 +227,7 @@ void QuickstartSTable()
 	// Layer[ 0] { { "state", class std::basic_string<char,struct std::char_traits<char>,class std::allocator<char> > } { "county", class std::basic_string<char,struct std::char_traits<char>,class std::allocator<char> > } { "city", class std::basic_string<char,struct std::char_traits<char>,class std::allocator<char> > } { "population_density", double } }
 
 	auto [estate, ecounty, ecity, epopulation_density] = extracted.GetPlaceholders("state"_fld, "county"_fld, "city"_fld, "population_density"_fld);
-	extracted | Show(estate, if_(len(ecounty) >= 16, substr(ecounty, 0, 13) + "...", ecounty), ecity, epopulation_density);
+	extracted | Show(estate, if_(len(ecounty) >= (size_t)16, substr(ecounty, 0, 13) + "...", ecounty), ecity, epopulation_density);
 
 	//If no _fld literal names are used, the result of Extract will be a DTree, because the tree structure is not determined statically.
 	adapt::DTree extracted_dtree = usa | Filter(city_area > 500.) | Extract(city.named("city"), (city_population / city_area).named("population_density"));
